added confirm and cancel section in Cart view controller fixed updating dishes functionality

+++ CustomerApp/Models/DishToOrder.swift
@@ -11,5 +11,5 @@ class DishToOrder{
     var dishId:Int?
     var dishName:String?
     var dishQuantity:Int?
-    var dishPrice:Int?
+    var dishPrice:Double?
 }
